+++ src/gui/imgui/CMakeLists.txt
@@ -1,5 +1,7 @@
 add_library(imgui.window window.cc)
-target_link_libraries(imgui.window)
+target_link_libraries(imgui.window
+	${IMGUI_LIBRARIES}
+)
 
 add_library(imgui.object_inspector object_inspector.cc)
 target_link_libraries(imgui.object_inspector
@@ -13,15 +15,8 @@ target_link_libraries(imgui.scene_explorer
 	scene.scene_manager
 )
 
-add_library(imgui.gui_base gui_base.cc)
-target_link_libraries(imgui.gui_base
-	imgui.window
-	${IMGUI_LIBRARIES}
-)
-
 add_library(imgui.gui_manager gui_manager.cc)
 target_link_libraries(imgui.gui_manager
-	imgui.gui_base
 	imgui.scene_explorer
 	imgui.object_inspector
 	core.engine
@@ -32,28 +27,18 @@ target_link_libraries(imgui.gui_manager_test
 	imgui.gui_manager
 )
 
+add_library(imgui.context context.cc)
+target_link_libraries(imgui.context
+	imgui.gui_manager
+)
+
 add_library(imgui.gui_system gui_system.cc)
 target_link_libraries(imgui.gui_system
-	imgui.gui_manager
+	imgui.context
+	${IMGUI_LIBRARIES}
 )
 
 add_gtest(imgui.gui_system_test gui_system_test.cc)
 target_link_libraries(imgui.gui_system_test
 	imgui.gui_system
-)
-
-add_library(imgui.context context.cc)
-target_link_libraries(imgui.context
-	imgui.gui_system
-	imgui.gui_manager
-)
-
-add_library(imgui.app app.cc)
-target_link_libraries(imgui.app
-	imgui.context
-)
-
-add_gtest(imgui.app_test app_test.cc)
-target_link_libraries(imgui.app_test
-	imgui.app
 )
