--- src/gui/imgui/app_test.cc
@@ -1,14 +1,0 @@
-#include <gtest/gtest.h>
-
-#include "src/gui/imgui/app.h"
-
-namespace kuro {
-
-namespace gui {
-
-// Demonstrate some basic assertions.
-TEST(AppTest, BasicAssertions) { App app; }
-
-}  // namespace gui
-
-}  // namespace kuro
