+++ src/gui/imgui/context.cc
@@ -10,17 +10,14 @@ namespace kuro {
 namespace gui {
 
 Context::Context() {
-  auto engine = Register<Engine>();
+  std::shared_ptr<Engine> engine = Engine::Instance();
+  Register<Engine>(engine);
 
   auto gui_store = Register<GuiStore>();
 
   auto gui_manager = Register<GuiManager>();
   gui_manager->RegisterEngine(engine);
   gui_manager->RegisterGuiStore(gui_store);
-
-  auto gui_system = Register<GuiSystem>();
-  gui_system->RegisterEngine(engine);
-  gui_system->RegisterGuiManager(gui_manager);
 }
 
 }  // namespace gui
