+++ src/gui/imgui/gui_system.cc
@@ -26,7 +26,11 @@ static void glfw_error_callback(int error, const char *description) {
   fprintf(stderr, "Glfw Error %d: %s\n", error, description);
 }
 
-GuiSystem::GuiSystem() {}
+GuiSystem::GuiSystem() {
+  context_ = std::make_shared<Context>();
+  engine_ = context_->Resolve<Engine>();
+  gui_manager_ = context_->Resolve<GuiManager>();
+}
 
 int GuiSystem::Init() {
   const unsigned int window_width = 3600, window_height = 1800;
@@ -118,7 +122,22 @@ int GuiSystem::Run() {
     glClearColor(0, 0, 0, 1);
     glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 
+    // Start the Dear ImGui frame
+    ImGui_ImplOpenGL3_NewFrame();
+    ImGui_ImplGlfw_NewFrame();
+    ImGui::NewFrame();
+
+    // 1. Show the big demo window (Most of the sample code is in
+    // ImGui::ShowDemoWindow()! You can browse its code to learn more about
+    // Dear ImGui!).
+    bool show_demo_window = true;
+    ImGui::ShowDemoWindow(&show_demo_window);
+
     Loop();
+
+    // Rendering
+    ImGui::Render();
+    ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
 
     glfwSwapBuffers(gl_window_);
   }
@@ -132,7 +151,11 @@ void GuiSystem::Loop() {
   gui_manager_->Draw();
 }
 
-GuiSystem::~GuiSystem() {}
+GuiSystem::~GuiSystem() {
+  ImGui_ImplOpenGL3_Shutdown();
+  ImGui_ImplGlfw_Shutdown();
+  ImGui::DestroyContext();
+}
 }  // namespace gui
 
 }  // namespace kuro
