+++ src/main.cc
@@ -1,9 +1,9 @@
 #include <iostream>
 
-#include "src/gui/imgui/app.h"
+#include "src/gui/imgui/gui_system.h"
 
 int main(int, char **) {
-  kuro::gui::App app;
-  app.Run();
+  kuro::gui::GuiSystem gui_system;
+  gui_system.Run();
   return 0;
 }
